number of tiles changed to 20
