+++ src/features/MainPage/popularMoviesSaga.js
@@ -8,20 +8,7 @@ import {
 
 function* fetchPopularMoviesHandler({ payload: pageNumber }) {
   try {
-    const numberOfPagesToGetFromApi = 2;
-    const highestNumberOfPageToGetFromApi =
-      pageNumber + numberOfPagesToGetFromApi - 1;
-    let pageNumberForApi = pageNumber;
-    let popularMoviesData = [];
-
-    for (
-      pageNumberForApi;
-      pageNumberForApi <= highestNumberOfPageToGetFromApi;
-      pageNumberForApi++
-    ) {
-      let dataFromApi = yield call(getPopularMovies, pageNumberForApi);
-      popularMoviesData = popularMoviesData.concat(dataFromApi);
-    }
+    const popularMoviesData = yield call(getPopularMovies, pageNumber);
 
     yield put(fetchPopularMoviesSuccess(popularMoviesData));
   } catch (error) {
